gedie_chat/requirements.txt
# Core – Telegram Bot
python-telegram-bot==20.7

# Banco de dados
SQLAlchemy==2.0.29
PyMySQL==1.1.1                  #  ← driver usado no URI mysql+pymysql://
# Se preferir mysqlconnector, troque o URI e remova PyMySQL
# mysql-connector-python==8.4.0

# Web (Flask dashboard)
Flask==3.0.3
Werkzeug==3.0.1                 # (vem com Flask)

# IA Gemini
google-generativeai==0.5.4
Pillow==10.3.0                  # gerar imagens de teste

# Config & utilidades
python-decouple==3.8            # leitura de .env
python-dotenv==1.0.1
loguru==0.7.2

# Testes
pytest==8.2.0
pytest-asyncio==0.23.7
pytest-cov==5.0.0
